requirements.txt
# Google Cloud & AI
google-auth>=2.26.0
google-cloud-aiplatform>=1.70.0
google-generativeai>=0.8.0

# API Clients
requests>=2.31.0
elevenlabs>=1.42.0
openai>=1.58.0
anthropic>=0.40.0
fal-client>=0.5.0
replicate>=0.32.0

# Data & Validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# Web & API Server
fastapi>=0.109.0
uvicorn>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Video Processing
ffmpeg-python>=0.2.3

# Utilities
Pillow>=10.2.0
